sample flask ms structure/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.3
importlib-metadata==6.11.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
Mako==1.3.2
MarkupSafe==2.1.5
pycparser==2.21
PyJWT==2.8.0
pyrsistent==0.20.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
Werkzeug==3.0.1
zipp==3.17.0
openpyxl==3.1.2
python-docx==1.1.0
pymysql==1.1.0
python-pptx==0.6.23
langchain==0.1.7
chromadb==0.4.22
openai==1.12.0
